requirements.txt
opencv-python
numpy
pillow
transformers
torch
